+++ config/debug.yaml
@@ -1,8 +1,8 @@
 options:
   trainer_options:
-    video_log_frequency: 3
+    video_log_frequency: 1
     log_frequency: 1
-    max_video_length: 40
+    max_video_length: 100
     eval_interaction_steps: 20
     eval_episode_length: 20
     training_logger_params:
@@ -10,6 +10,6 @@ options:
   seed: 1
   train_iterations: 5
   agent: mbrl_agent
-  environment: MbrlInvertedPendulum-v2
+  environment: MbrlSafexp-PointGoal1-v0
