+++ simba/environment_utils/__init__.py
@@ -1,7 +1,2 @@
-from gym.envs.registration import register
-from simba.environment_utils.safety_gym_scoring import SafetyGymStateScorer
-
-register(
-    id='MbrlInvertedPendulum-v2',
-    entry_point='simba.environment_utils.inverted_pendulum:MbrlInvertedPendulumEnv'
-)
+from simba.environment_utils.safety_gym import MbrlSafetyGym
+from simba.environment_utils.inverted_pendulum import MbrlInvertedPendulum
